refactor(settings): refactored the individual settings containers into the main container
re #1481
